Move WorkspaceStats.TAGS to a service, fixes #76002
